src/main/java/com/liuwjg/service/impl/CategoryServiceImpl.java
package com.liuwjg.service.impl;

import com.liuwjg.dataobject.ProductCategory;
import com.liuwjg.repository.ProductCategoryRepository;
import com.liuwjg.service.CategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;
@Service
public class CategoryServiceImpl implements CategoryService {

    @Autowired
    ProductCategoryRepository repository;
    @Override
    public ProductCategory findById(Integer id) {
        return repository.findById(1).get();
    }

    @Override
    public List<ProductCategory> findAll() {
        return repository.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(List<Integer> categoryTypeList) {
        return repository.findByCategoryTypeIn(categoryTypeList);
    }

    @Override
    public ProductCategory save(ProductCategory productCategory) {
        return repository.save(productCategory);
    }
}
